Completed Accounts Screen UI, managed navigation compatibility with other components
Accounts Screen UI should be complete, but still lacks backend logic and connection to firebase. Should be working system.

Navigation Builder is reconfigured to be compatible with Top App Bar, so that arguments can be passed into top app bar to edit title display. Also, arguments are now used to pass strings into Budget Item screen and Account Transactions Screen. To be converted to doc ID when backend model is approved.

Also, onScreenChange is removed from AppState and moved to NavHostBuilder. This is because the builder is now passing arguments, and retrieving Screen enum from backstack is no longer possible without extracting a substring.